Output ppmw of species in reservoir
For the mantle reservoirs, the ppmw is relative to the mass of silicates (either molten or solid). For the atmosphere, the ppmw is relative to the total mass of the atmosphere (which itself is only composed of gas species).  The ppmw is output for both gas species as well as the element breakdowns.