Perturb initial guess
If the initial guess fails, randomly perturb the initial guess by plus or minus 'log10_perturb', which defaults to one (log10 unit). Then try to find a solution. Repeat a maximum of 'max_attempts' and then either raise an Exception if 'errors'=raise or alternatively skip the calculation if 'errors'=ignore.