Specify axes specifications for each species
The data range can vary substantially for species, depending on the species, its solubility, and whether we desire to show some or all species using the same data ranges for direct visual comparison. This allows the use to specify a value in a dictionary to control the formatting of the axes, or specify 'None' to allow for an automatic adjustment.