Water condensation test
Now that H2O can exist as both a gas and a liquid, it's necessary to honour the phase as a suffix to the formula, otherwise there is no unique naming convention.  Therefore, this will break reading in data that was previously generated, because now the naming convention for the species output is _g for gas phases, _l for liquid phases, and _cr for solid phases (abiding to the JANAF naming convention).